Three Planes v3.0
Intersection of three planes is completed
When an solution is returned, first the object with the solution is returned then the name corresponding to the image is returned in a tuple
***Reduce Function still giving errors when working with planes***
Maybe try to make it recursive? Might be because it's not taking into account D, or when values are 0 how to reduce the other ones?